refactoring media server catalog
