+++ shared/src/model/config/input.rs
@@ -3,13 +3,13 @@ use crate::{
     check_input_connections, check_input_credentials,
     error::TuliproxError,
     foundation::{get_filter, Filter},
-    model::{EpgConfigDto, PatternTemplate},
+    model::{config::media_server_catalog::MediaServerInputConfigDto, EpgConfigDto, PatternTemplate},
     utils::{
         arc_str_serde, arc_str_vec_serde, default_as_true, default_probe_delay_secs, default_probe_live_interval,
         default_resolve_background, default_resolve_delay_secs, default_xtream_live_stream_use_prefix,
         deserialize_timestamp, get_credentials_from_url_str, get_trimmed_string, is_blank_optional_string,
         is_default_probe_delay_secs, is_default_probe_live_interval, is_default_resolve_delay_secs, is_false,
-        is_non_blank_optional_string, is_true, is_zero_i16, is_zero_u16, parse_duration_seconds,
+        is_true, is_zero_i16, is_zero_u16, parse_duration_seconds,
         parse_provider_scheme_url_parts, sanitize_sensitive_info, serialize_option_vec_flow_map_items, trim_last_slash,
         Internable, BATCH_SCHEME_PREFIX, PROVIDER_SCHEME_PREFIX,
     },
@@ -333,286 +333,6 @@ impl ConfigInputOptionsDto {
             self.t_probe_filter = Some(get_filter(raw_filter, templates)?);
         }
         Ok(())
-    }
-}
-
-pub const fn default_media_server_catalog_page_size() -> u16 { 100 }
-pub const fn default_media_server_catalog_request_delay_ms() -> u64 { 250 }
-pub const fn is_default_media_server_catalog_page_size(value: &u16) -> bool {
-    *value == default_media_server_catalog_page_size()
-}
-pub const fn is_default_media_server_catalog_request_delay_ms(value: &u64) -> bool {
-    *value == default_media_server_catalog_request_delay_ms()
-}
-
-#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize, Sequence, PartialEq, Eq, Default)]
-pub enum MediaServerCatalogRefreshModeDto {
-    #[serde(rename = "manual")]
-    #[default]
-    Manual,
-    #[serde(rename = "scheduled")]
-    Scheduled,
-}
-
-pub fn is_default_media_server_catalog_refresh_mode(value: &MediaServerCatalogRefreshModeDto) -> bool {
-    *value == MediaServerCatalogRefreshModeDto::default()
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct MediaServerCatalogConfigDto {
-    #[serde(default, skip_serializing_if = "is_default_media_server_catalog_refresh_mode")]
-    pub refresh_mode: MediaServerCatalogRefreshModeDto,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub refresh_on_startup: bool,
-    #[serde(
-        default = "default_media_server_catalog_page_size",
-        skip_serializing_if = "is_default_media_server_catalog_page_size"
-    )]
-    pub page_size: u16,
-    #[serde(
-        default = "default_media_server_catalog_request_delay_ms",
-        skip_serializing_if = "is_default_media_server_catalog_request_delay_ms"
-    )]
-    pub request_delay_ms: u64,
-    #[serde(default = "default_as_true", skip_serializing_if = "is_true")]
-    pub include_media_sources: bool,
-    #[serde(default, alias = "include_file_paths", skip_serializing_if = "is_false")]
-    pub include_paths: bool,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub include_user_state: bool,
-}
-
-impl Default for MediaServerCatalogConfigDto {
-    fn default() -> Self {
-        Self {
-            refresh_mode: MediaServerCatalogRefreshModeDto::default(),
-            refresh_on_startup: false,
-            page_size: default_media_server_catalog_page_size(),
-            request_delay_ms: default_media_server_catalog_request_delay_ms(),
-            include_media_sources: default_as_true(),
-            include_paths: false,
-            include_user_state: false,
-        }
-    }
-}
-
-impl MediaServerCatalogConfigDto {
-    pub fn is_default(&self) -> bool { self == &Self::default() }
-
-    pub fn prepare(&self, input_name: &Arc<str>) -> Result<(), TuliproxError> {
-        if self.page_size == 0 {
-            return Err(TuliproxError::ConfigInput(format!(
-                "media server catalog page_size must be greater than zero (input: {input_name})"
-            )));
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize, Sequence, PartialEq, Eq, Default)]
-pub enum MediaServerPlaybackInfoPolicyDto {
-    #[serde(rename = "on_demand")]
-    #[default]
-    OnDemand,
-    #[serde(rename = "disabled")]
-    Disabled,
-}
-
-pub fn is_default_media_server_playback_info_policy(value: &MediaServerPlaybackInfoPolicyDto) -> bool {
-    *value == MediaServerPlaybackInfoPolicyDto::default()
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct MediaServerPlaybackConfigDto {
-    #[serde(default, skip_serializing_if = "is_default_media_server_playback_info_policy")]
-    pub playback_info_policy: MediaServerPlaybackInfoPolicyDto,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub preflight_streams: bool,
-    #[serde(default = "default_as_true", skip_serializing_if = "is_true")]
-    pub direct_play_only: bool,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub allow_transcode: bool,
-}
-
-impl Default for MediaServerPlaybackConfigDto {
-    fn default() -> Self {
-        Self {
-            playback_info_policy: MediaServerPlaybackInfoPolicyDto::default(),
-            preflight_streams: false,
-            direct_play_only: default_as_true(),
-            allow_transcode: false,
-        }
-    }
-}
-
-impl MediaServerPlaybackConfigDto {
-    pub fn is_default(&self) -> bool { self == &Self::default() }
-}
-
-#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize, Sequence, PartialEq, Eq, Default)]
-pub enum MediaServerImagePolicyDto {
-    #[serde(rename = "proxy_on_demand")]
-    #[default]
-    ProxyOnDemand,
-    #[serde(rename = "disabled")]
-    Disabled,
-}
-
-pub fn is_default_media_server_image_policy(value: &MediaServerImagePolicyDto) -> bool {
-    *value == MediaServerImagePolicyDto::default()
-}
-
-#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize, Sequence, PartialEq, Eq)]
-pub enum MediaServerLibraryKindDto {
-    #[serde(rename = "movies")]
-    Movies,
-    #[serde(rename = "tvshows", alias = "shows", alias = "series")]
-    TvShows,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct MediaServerLibrarySelectorDetailsDto {
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub id: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub key: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub kind: Option<MediaServerLibraryKindDto>,
-}
-
-impl MediaServerLibrarySelectorDetailsDto {
-    fn prepare(&mut self) {
-        self.id = get_trimmed_string(self.id.as_deref());
-        self.key = get_trimmed_string(self.key.as_deref());
-        self.name = get_trimmed_string(self.name.as_deref());
-    }
-
-    fn is_empty(&self) -> bool {
-        self.id.as_ref().is_none_or(|s| s.trim().is_empty())
-            && self.key.as_ref().is_none_or(|s| s.trim().is_empty())
-            && self.name.as_ref().is_none_or(|s| s.trim().is_empty())
-            && self.kind.is_none()
-    }
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum MediaServerLibrarySelectorDto {
-    Name(String),
-    Detailed(MediaServerLibrarySelectorDetailsDto),
-}
-
-impl MediaServerLibrarySelectorDto {
-    fn prepare(&mut self) {
-        match self {
-            Self::Name(name) => *name = name.trim().to_string(),
-            Self::Detailed(details) => details.prepare(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        match self {
-            Self::Name(name) => name.trim().is_empty(),
-            Self::Detailed(details) => details.is_empty(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct MediaServerInputConfigDto {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub libraries: Vec<MediaServerLibrarySelectorDto>,
-    #[serde(default, skip_serializing_if = "MediaServerCatalogConfigDto::is_default")]
-    pub catalog: MediaServerCatalogConfigDto,
-    #[serde(default, skip_serializing_if = "MediaServerPlaybackConfigDto::is_default")]
-    pub playback: MediaServerPlaybackConfigDto,
-    #[serde(default, skip_serializing_if = "is_default_media_server_image_policy")]
-    pub image_policy: MediaServerImagePolicyDto,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub token: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub api_key: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub user_id: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub account_token: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub server_id: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub machine_id: Option<String>,
-    #[serde(default, skip_serializing_if = "is_blank_optional_string")]
-    pub server_name: Option<String>,
-    #[serde(default = "default_as_true", skip_serializing_if = "is_true")]
-    pub prefer_https: bool,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub allow_relay: bool,
-}
-
-impl Default for MediaServerInputConfigDto {
-    fn default() -> Self {
-        Self {
-            libraries: Vec::new(),
-            catalog: MediaServerCatalogConfigDto::default(),
-            playback: MediaServerPlaybackConfigDto::default(),
-            image_policy: MediaServerImagePolicyDto::default(),
-            token: None,
-            api_key: None,
-            user_id: None,
-            account_token: None,
-            server_id: None,
-            machine_id: None,
-            server_name: None,
-            prefer_https: default_as_true(),
-            allow_relay: false,
-        }
-    }
-}
-
-impl MediaServerInputConfigDto {
-    pub fn normalize(&mut self) {
-        self.token = get_trimmed_string(self.token.as_deref());
-        self.api_key = get_trimmed_string(self.api_key.as_deref());
-        self.user_id = get_trimmed_string(self.user_id.as_deref());
-        self.account_token = get_trimmed_string(self.account_token.as_deref());
-        self.server_id = get_trimmed_string(self.server_id.as_deref());
-        self.machine_id = get_trimmed_string(self.machine_id.as_deref());
-        self.server_name = get_trimmed_string(self.server_name.as_deref());
-
-        for library in &mut self.libraries {
-            library.prepare();
-        }
-    }
-
-    pub fn prepare(&mut self, input_name: &Arc<str>) -> Result<(), TuliproxError> {
-        self.normalize();
-        self.catalog.prepare(input_name)?;
-
-        if self.libraries.iter().any(MediaServerLibrarySelectorDto::is_empty) {
-            return Err(TuliproxError::ConfigInput(format!(
-                "media_server library selectors must not be empty (input: {input_name})"
-            )));
-        }
-        Ok(())
-    }
-
-    pub fn has_any_emby_jellyfin_auth(&self) -> bool {
-        is_non_blank_optional_string(&self.token) || is_non_blank_optional_string(&self.api_key)
-    }
-
-    pub fn has_any_plex_token(&self) -> bool {
-        is_non_blank_optional_string(&self.account_token) || is_non_blank_optional_string(&self.token)
-    }
-
-    pub fn has_plex_server_selector(&self) -> bool {
-        is_non_blank_optional_string(&self.server_id)
-            || is_non_blank_optional_string(&self.machine_id)
-            || is_non_blank_optional_string(&self.server_name)
     }
 }
 
@@ -1426,272 +1146,8 @@ mod tests {
         ConfigInputDto { name: "test_input".intern(), ..ConfigInputDto::default() }
     }
 
-    fn media_server_config_with_library() -> MediaServerInputConfigDto {
-        MediaServerInputConfigDto {
-            libraries: vec![MediaServerLibrarySelectorDto::Name("Movies".to_string())],
-            ..MediaServerInputConfigDto::default()
-        }
-    }
-
-    #[test]
-    fn prepare_rejects_blank_media_server_credentials_and_selectors() {
-        let mut emby = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            url: "https://media.example.invalid".to_string(),
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some("   ".to_string()),
-                api_key: Some("".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-        let err = prepare_dto(&mut emby).expect_err("blank token/api_key should be rejected");
-        assert!(err.to_string().contains("requires media_server token/api_key"));
-
-        let mut plex = ConfigInputDto {
-            name: "plex_media_server".intern(),
-            input_type: InputType::Plex,
-            media_server: Some(MediaServerInputConfigDto {
-                account_token: Some("   ".to_string()),
-                server_id: Some("   ".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-        let err = prepare_dto(&mut plex).expect_err("blank plex token should be rejected");
-        assert!(err.to_string().contains("requires media_server.account_token or media_server.token"));
-    }
-
-    #[test]
-    fn prepare_accepts_media_server_max_connections_as_stream_limit() {
-        let mut dto = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            url: "https://media.example.invalid".to_string(),
-            max_connections: 1,
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some("token-value".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        prepare_dto(&mut dto).expect("media_server inputs reuse max_connections stream-limit semantics");
-    }
-
     fn prepare_dto(dto: &mut ConfigInputDto) -> Result<u16, TuliproxError> {
         dto.prepare(0, false, &HashSet::new(), None)
-    }
-
-    #[test]
-    fn media_server_defaults_are_conservative() {
-        let media_server = MediaServerInputConfigDto::default();
-
-        assert_eq!(media_server.catalog.page_size, 100);
-        assert_eq!(media_server.catalog.request_delay_ms, 250);
-        assert!(media_server.catalog.include_media_sources);
-        assert!(!media_server.catalog.include_paths);
-        assert!(!media_server.catalog.include_user_state);
-        assert!(!media_server.catalog.refresh_on_startup);
-        assert!(media_server.playback.direct_play_only);
-        assert!(!media_server.playback.allow_transcode);
-        assert!(!media_server.playback.preflight_streams);
-        assert_eq!(media_server.image_policy, MediaServerImagePolicyDto::ProxyOnDemand);
-        assert!(!media_server.allow_relay);
-    }
-
-    #[test]
-    fn media_server_enrichment_block_is_not_part_of_schema() {
-        let err = serde_json::from_str::<MediaServerInputConfigDto>(
-            r#"{"libraries":["Movies"],"enrichment":{"ffprobe":true,"tmdb_lookup":true,"fetch_images":true}}"#,
-        )
-        .expect_err("media_server.enrichment must not be accepted");
-
-        assert!(err.to_string().contains("unknown field `enrichment`"));
-    }
-
-    #[test]
-    fn prepare_accepts_emby_media_server_with_token_and_library() {
-        let mut dto = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            url: " https://media.example.invalid/ ".to_string(),
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some(" token-value ".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        prepare_dto(&mut dto).expect("emby media_server config should prepare");
-
-        assert_eq!(dto.url, "https://media.example.invalid/");
-        assert!(dto.input_type.is_media_server());
-        assert_eq!(
-            dto.media_server.as_ref().and_then(|media_server| media_server.token.as_deref()),
-            Some("token-value")
-        );
-    }
-
-    #[test]
-    fn prepare_rejects_media_server_without_media_server_block() {
-        let mut dto = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            url: "https://media.example.invalid".to_string(),
-            ..ConfigInputDto::default()
-        };
-
-        let err = prepare_dto(&mut dto).expect_err("media_server block should be mandatory");
-        assert!(err.to_string().contains("media_server configuration is mandatory"));
-    }
-
-    #[test]
-    fn prepare_allows_disabled_media_server_input_with_incomplete_config() {
-        let mut dto = ConfigInputDto {
-            name: "disabled_plex".intern(),
-            input_type: InputType::Plex,
-            enabled: false,
-            ..ConfigInputDto::default()
-        };
-
-        prepare_dto(&mut dto).expect("disabled media_server input should not require active playback/catalog config");
-        assert_eq!(dto.input_type, InputType::Plex);
-        assert!(!dto.enabled);
-    }
-
-    #[test]
-    fn prepare_normalizes_disabled_media_server_config_without_enforcing_invariants() {
-        let mut dto = ConfigInputDto {
-            name: "disabled_emby".intern(),
-            input_type: InputType::Emby,
-            enabled: false,
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some(" token-value ".to_string()),
-                libraries: vec![MediaServerLibrarySelectorDto::Name("   ".to_string())],
-                catalog: MediaServerCatalogConfigDto { page_size: 0, ..MediaServerCatalogConfigDto::default() },
-                ..MediaServerInputConfigDto::default()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        prepare_dto(&mut dto).expect("disabled media_server input can preserve incomplete config for later repair");
-        let media_server = dto.media_server.as_ref().expect("media_server config should be preserved");
-        assert_eq!(media_server.token.as_deref(), Some("token-value"));
-        assert!(media_server.libraries[0].is_empty());
-    }
-
-    #[test]
-    fn prepare_rejects_emby_media_server_without_input_url() {
-        let mut dto = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some("token-value".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        let err = prepare_dto(&mut dto).expect_err("emby media_server input should require a direct server URL");
-        assert!(err.to_string().contains("url is mandatory for input type emby"));
-    }
-
-    #[test]
-    fn prepare_rejects_media_server_provider_scheme_url() {
-        let mut dto = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            url: " provider://media-server ".to_string(),
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some("token-value".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        let err = prepare_dto(&mut dto).expect_err("media_server input must not use provider URLs");
-        assert!(err.to_string().contains("does not support batch:// or provider://"));
-    }
-
-    #[test]
-    fn prepare_rejects_media_server_staged_input() {
-        let mut dto = ConfigInputDto {
-            name: "emby_media_server".intern(),
-            input_type: InputType::Emby,
-            url: "https://media.example.invalid".to_string(),
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some("token-value".to_string()),
-                ..media_server_config_with_library()
-            }),
-            staged: Some(StagedInputDto { enabled: true, name: "staged".intern(), ..StagedInputDto::default() }),
-            ..ConfigInputDto::default()
-        };
-
-        let err = prepare_dto(&mut dto).expect_err("media_server input must reject staged config");
-        assert!(err.to_string().contains("does not support staged inputs"));
-    }
-
-    #[test]
-    fn prepare_rejects_plex_without_token_or_server_selector() {
-        let mut without_token = ConfigInputDto {
-            name: "plex_media_server".intern(),
-            input_type: InputType::Plex,
-            media_server: Some(MediaServerInputConfigDto {
-                machine_id: Some("machine".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-        let err = prepare_dto(&mut without_token).expect_err("plex token should be mandatory");
-        assert!(err.to_string().contains("requires media_server.account_token or media_server.token"));
-
-        let mut without_selector = ConfigInputDto {
-            name: "plex_media_server".intern(),
-            input_type: InputType::Plex,
-            media_server: Some(MediaServerInputConfigDto {
-                account_token: Some("token".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-        let err = prepare_dto(&mut without_selector).expect_err("plex server selector should be mandatory");
-        assert!(err.to_string().contains("requires a server selector"));
-    }
-
-    #[test]
-    fn prepare_accepts_plex_without_input_url_when_discovery_is_configured() {
-        let mut dto = ConfigInputDto {
-            name: "plex_media_server".intern(),
-            input_type: InputType::Plex,
-            media_server: Some(MediaServerInputConfigDto {
-                account_token: Some("token".to_string()),
-                machine_id: Some("machine".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        prepare_dto(&mut dto).expect("plex discovery config should not require input.url");
-        assert_eq!(dto.input_type, InputType::Plex);
-    }
-
-    #[test]
-    fn prepare_accepts_plex_media_server_with_direct_url_without_selector() {
-        let mut dto = ConfigInputDto {
-            name: "plex_media_server".intern(),
-            input_type: InputType::Plex,
-            url: "https://plex.example.invalid".to_string(),
-            media_server: Some(MediaServerInputConfigDto {
-                token: Some("token".to_string()),
-                ..media_server_config_with_library()
-            }),
-            ..ConfigInputDto::default()
-        };
-
-        prepare_dto(&mut dto).expect("direct Plex URL should not require MyPlex server selector");
-        assert_eq!(dto.input_type, InputType::Plex);
     }
 
     #[test]
